Feature/book 001 t3

+++ library-backend/src/main/java/com/library/repository/AuthorRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AuthorRepository extends JpaRepository<Author, Integer> {
+public interface AuthorRepository extends JpaRepository<Author, Long> {
     
     List<Author> findByNameContainingIgnoreCase(String name);
     
@@ -21,14 +21,14 @@ public interface AuthorRepository extends JpaRepository<Author, Integer> {
     List<Author> findByNationality(String nationality);
     
     @Query("SELECT DISTINCT a FROM Author a LEFT JOIN FETCH a.bookAuthors ba LEFT JOIN FETCH ba.book WHERE a.id = :id")
-    Optional<Author> findByIdWithBooks(@Param("id") Integer id);
+    Optional<Author> findByIdWithBooks(@Param("id") Long id);
     
     @Query("SELECT a FROM Author a WHERE LOWER(a.name) LIKE LOWER(CONCAT('%', :keyword, '%')) " +
            "OR LOWER(a.biography) LIKE LOWER(CONCAT('%', :keyword, '%'))")
     Page<Author> searchByKeyword(@Param("keyword") String keyword, Pageable pageable);
     
     @Query("SELECT COUNT(ba) FROM BookAuthor ba WHERE ba.author.id = :authorId")
-    Long countBooksByAuthorId(@Param("authorId") Integer authorId);
+    Long countBooksByAuthorId(@Param("authorId") Long authorId);
     
     @Query("SELECT DISTINCT a.nationality FROM Author a WHERE a.nationality IS NOT NULL")
     List<String> findAllNationalities();
